common-tap: Remove preescaped node/npm paths
Because how you escape differs depending on where you're using these things,
its better to leave that to the tests. Eg, if you're calling `node`
via `spawn` then it should be escaped with excapeExecPath, but if you're
calling it via `npm explore` then it needs to be escaped with `escapeArg`.